+++ go/vt/vttablet/tabletmanager/vreplication/table_plan_builder.go
@@ -19,11 +19,8 @@ package vreplication
 import (
 	"fmt"
 	"regexp"
-	"runtime/debug"
 	"sort"
 	"strings"
-
-	"vitess.io/vitess/go/vt/log"
 
 	"vitess.io/vitess/go/mysql/collations"
 	"vitess.io/vitess/go/sqltypes"
@@ -120,7 +117,6 @@ const (
 )
 
 func buildReplicatorPlan(source *binlogdatapb.BinlogSource, colInfoMap map[string][]*ColumnInfo, copyState map[string]*sqltypes.Result, stats *binlogplayer.Stats, collationEnv *collations.Environment, parser *sqlparser.Parser) (*ReplicatorPlan, error) {
-	log.Infof("In buildReplicatorPlan %s", debug.Stack())
 	filter := source.Filter
 	plan := &ReplicatorPlan{
 		VStreamFilter: &binlogdatapb.Filter{FieldEventMode: filter.FieldEventMode},
@@ -157,7 +153,6 @@ func buildReplicatorPlan(source *binlogdatapb.BinlogSource, colInfoMap map[strin
 			continue
 		}
 		if dup, ok := plan.TablePlans[tablePlan.SendRule.Match]; ok {
-			log.Infof("more than one target for source table %s: %s and %s", tablePlan.SendRule.Match, dup.TargetName, tableName)
 			return nil, fmt.Errorf("more than one target for source table %s: %s and %s", tablePlan.SendRule.Match, dup.TargetName, tableName)
 		}
 		plan.VStreamFilter.Rules = append(plan.VStreamFilter.Rules, tablePlan.SendRule)
@@ -165,224 +160,6 @@ func buildReplicatorPlan(source *binlogdatapb.BinlogSource, colInfoMap map[strin
 		plan.TablePlans[tablePlan.SendRule.Match] = tablePlan
 	}
 	return plan, nil
-}
-
-/*
-Need to create insert for target table
-*/
-func buildReplicatorPlanForJoin(source *binlogdatapb.BinlogSource, colInfoMap map[string][]*ColumnInfo,
-	copyState map[string]*sqltypes.Result, stats *binlogplayer.Stats, collationEnv *collations.Environment,
-	parser *sqlparser.Parser, dbClient *vdbClient, copy bool) (*ReplicatorPlan, error) {
-
-	joinTables, err := vttablet.GetJoinedTables(parser, source.Filter.Rules[0].Filter)
-	if err != nil {
-		return nil, err
-	}
-	if len(joinTables) == 0 {
-		return buildReplicatorPlan(source, colInfoMap, copyState, stats, collationEnv, parser)
-	}
-	log.Infof("In buildReplicatorPlanForJoin with join tables %v", joinTables)
-	filter := source.Filter
-	joinPlan := &ReplicatorJoinPlan{
-		Tables: joinTables,
-	}
-	plan := &ReplicatorPlan{
-		VStreamFilter: &binlogdatapb.Filter{FieldEventMode: filter.FieldEventMode},
-		TargetTables:  make(map[string]*TablePlan),
-		TablePlans:    make(map[string]*TablePlan),
-		ColInfoMap:    colInfoMap,
-		stats:         stats,
-		Source:        source,
-		collationEnv:  collationEnv,
-		joinPlan:      joinPlan,
-	}
-	// FIXME: check copy state?
-
-	plan.VStreamFilter = source.Filter.CloneVT()
-	view := source.Filter.Rules[0].Match
-	joinPlan.ViewTableName = view
-	joinPlan.BaseTableName = joinTables[0]
-	query := source.Filter.Rules[0].Filter
-	log.Infof("View %s, query %s", view, query)
-	tablePlan := &TablePlan{
-		TargetName: view,
-		SendRule:   source.Filter.Rules[0],
-	}
-	colInfos, ok := colInfoMap[view]
-	if !ok {
-		return nil, fmt.Errorf("table %s not found in schema", view)
-	}
-
-	statement, err := parser.Parse(query)
-	if err != nil {
-		return nil, err
-	}
-	sel, ok := statement.(*sqlparser.Select)
-	if !ok {
-		return nil, fmt.Errorf("unsupported non-select statement")
-	}
-
-	tpb := &tablePlanBuilder{
-		name: sqlparser.NewIdentifierCS(view),
-		sendSelect: &sqlparser.Select{
-			From:  sel.From,
-			Where: sel.Where,
-		},
-		lastpk:       nil,
-		colInfos:     colInfos,
-		stats:        stats,
-		source:       source,
-		collationEnv: collationEnv,
-		viewMode:     true,
-	}
-	err = tpb.analyzeExprs(sel.SelectExprs)
-	if err != nil {
-		return nil, err
-	}
-	tpb.pkCols = append(tpb.pkCols, tpb.colExprs[0])
-	tablePlan.PKReferences = []string{tpb.colExprs[0].colName.String()}
-	tableColumns := make(map[string][]*ViewColumn)
-
-	for _, selExpr := range sel.SelectExprs {
-		aliasExpr, ok := selExpr.(*sqlparser.AliasedExpr)
-		if !ok {
-			return nil, fmt.Errorf("unsupported expression %s", sqlparser.String(selExpr))
-		}
-		colName, ok := aliasExpr.Expr.(*sqlparser.ColName)
-		if !ok {
-			return nil, fmt.Errorf("unsupported expression %s", sqlparser.String(aliasExpr.Expr))
-		}
-		if colName.Qualifier.IsEmpty() {
-			return nil, fmt.Errorf("every column needs to have a table qualifier %s", colName)
-		}
-		tableName := colName.Qualifier.Name.String()
-		columnName := colName.Name.String()
-		aliasName := columnName
-		if !aliasExpr.As.IsEmpty() {
-			aliasName = aliasExpr.As.String()
-		}
-		tableColumns[tableName] = append(tableColumns[tableName], &ViewColumn{
-			SourceColumnName: columnName,
-			ViewColumnName:   aliasName,
-		})
-	}
-	for table, cols := range tableColumns {
-		log.Infof("table %s, column %v", table, cols)
-	}
-
-	if view == "" {
-		return nil, fmt.Errorf("view name is empty")
-	}
-	qr, err := dbClient.ExecuteFetch(fmt.Sprintf("select * from %s limit 1", view), 1)
-	if err != nil {
-		return nil, err
-
-	}
-	tablePlan.Fields = qr.Fields
-	insert := generateInsertForJoin(view, query, tablePlan.Fields)
-	log.Infof("Insert for Join is %+q", insert)
-	updates := generateUpdatesForJoin(view, tableColumns)
-	deletes := generateDeletesForJoin(view, tableColumns)
-	tablePlan.JoinPlan = &TableJoinPlan{
-		BaseTableName: joinTables[0],
-		TableColumns:  &tableColumns,
-		Insert:        insert,
-		Updates:       updates,
-		Deletes:       deletes,
-	}
-	if !copy {
-		plan.VStreamFilter.Rules[0].Match = fmt.Sprintf("/\\b(%s)\\b", strings.Join(plan.joinPlan.Tables, "|"))
-	} else {
-		plan.VStreamFilter.Rules[0].Filter = fmt.Sprintf("/*vt+ view=%s */ %s", view, plan.VStreamFilter.Rules[0].Filter)
-	}
-	plan.TablePlans[view] = tablePlan
-	plan.TargetTables[view] = tablePlan
-	log.Infof("Added table plan for view %s: %v, insert %v, updates %+v, deletes %+v", view, tablePlan, insert, updates, deletes)
-	for _, tableName := range joinTables {
-		rule := &binlogdatapb.Rule{
-			Match: tableName,
-		}
-		tablePlan, err := buildTablePlan(tableName, rule, colInfos, nil, stats, source, collationEnv, parser)
-		if err != nil {
-			return nil, err
-		}
-		plan.TablePlans[tableName] = tablePlan
-	}
-	return plan, nil
-}
-
-type ViewColumn struct {
-	SourceColumnName string
-	ViewColumnName   string
-}
-
-func generateInsertForJoin(view string, viewQuery string, fields []*querypb.Field) *sqlparser.ParsedQuery {
-	bvf := &bindvarFormatter{}
-	buf := sqlparser.NewTrackedBuffer(bvf.formatter)
-	bvf.mode = bvAfter
-	buf.Myprintf("insert into %v (", sqlparser.NewIdentifierCS(view))
-	separator := ""
-	for _, field := range fields {
-		buf.Myprintf("%s%v", separator, sqlparser.NewIdentifierCI(field.Name))
-		separator = ","
-	}
-	buf.Myprintf(") ")
-	buf.Myprintf("%s", viewQuery)
-	buf.Myprintf(" where oid = ")
-	buf.Myprintf("%v", sqlparser.NewColName("oid"))
-	return buf.ParsedQuery()
-}
-
-func generateUpdatesForJoin(view string, viewColumns map[string][]*ViewColumn) map[string]*sqlparser.ParsedQuery {
-	updates := map[string]*sqlparser.ParsedQuery{}
-	for table, cols := range viewColumns {
-		bvf := &bindvarFormatter{}
-		buf := sqlparser.NewTrackedBuffer(bvf.formatter)
-		bvf.mode = bvAfter
-		buf.Myprintf("update %v set ", sqlparser.NewIdentifierCS(view))
-		separator := ""
-		for i, col := range cols {
-			if i == 0 {
-				continue
-			}
-			buf.Myprintf("%s%s = ", separator, sqlparser.NewIdentifierCI(col.ViewColumnName).CompliantName())
-			buf.Myprintf("%v", sqlparser.NewColName(col.ViewColumnName))
-			separator = ", "
-		}
-		col := cols[0]
-		bvf.mode = bvBefore
-		buf.Myprintf(" where %s = ", sqlparser.NewIdentifierCI(col.ViewColumnName).String())
-		buf.Myprintf("%v", sqlparser.NewColName(col.ViewColumnName))
-		updates[table] = buf.ParsedQuery()
-		log.Infof("Update for table %s is %s, bindLocations %d", table, buf.String(), len(updates[table].BindLocations()))
-	}
-	return updates
-}
-
-func generateDeletesForJoin(view string, viewColumns map[string][]*ViewColumn) map[string]*sqlparser.ParsedQuery {
-	deletes := map[string]*sqlparser.ParsedQuery{}
-	for table, cols := range viewColumns {
-		bvf := &bindvarFormatter{}
-		buf := sqlparser.NewTrackedBuffer(bvf.formatter)
-		bvf.mode = bvBefore
-		buf.Myprintf("delete from %v where ", sqlparser.NewIdentifierCS(view))
-		separator := ""
-		for i, col := range cols {
-			if i == 0 {
-				continue
-			}
-			buf.Myprintf("%s%s = ", separator, sqlparser.NewIdentifierCI(col.ViewColumnName).CompliantName())
-			buf.Myprintf("%v", sqlparser.NewColName(col.ViewColumnName))
-			separator = " and "
-		}
-		deletes[table] = buf.ParsedQuery()
-		log.Infof("Delete for table %s is %s, bindLocations %d", table, buf.String(), len(deletes[table].BindLocations()))
-	}
-	return deletes
-}
-
-func (vc *ViewColumn) String() string {
-	return fmt.Sprintf("SourceColumnName %s, ViewColumnName %s", vc.SourceColumnName, vc.ViewColumnName)
 }
 
 // MatchTable is similar to tableMatches and buildPlan defined in vstreamer/planbuilder.go.
